+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 ### Added
 - Sidebar now shows tick count until daily bonuses reset
 - Dominions that remain inactive and under protection for three days will be moved to realm 0
+- Dominions over 600 acres may no longer leave protection with less than the minimum defense
 
 ### Changed
 - Default draft rate set to 35%